requirements.txt
# Python dependencies
flask
google-genai
google-api-python-client
google-auth
gunicorn